+++ packages/grid/src/features/standard-table/hooks/UseCellBackground.ts
@@ -9,20 +9,24 @@ const getBackgroundColor = <TItem>(
   backgroundResolver: BackgroundResolver<TItem> | undefined,
   item: TItem,
   background: string | undefined
-) => (backgroundResolver ? backgroundResolver(item) : background) ?? "inherit";
+): string | undefined =>
+  backgroundResolver ? backgroundResolver(item) : background;
 
 const useBackground = <TItem>(
   backgroundResolver: BackgroundResolver<TItem> | undefined,
   item: TItem,
   background: string | undefined
-) =>
+): string | undefined =>
   useMemo(() => getBackgroundColor(backgroundResolver, item, background), [
     backgroundResolver,
     item,
     background,
   ]);
 
-export const useCellBackgroundByColumnId = <T>(columnId: string, item: T) => {
+export const useCellBackgroundByColumnId = <T>(
+  columnId: string,
+  item: T
+): string | undefined => {
   const { background, backgroundResolver } = useColumnConfigById(columnId);
   return useBackground(backgroundResolver, item, background);
 };
@@ -30,7 +34,7 @@ export const useCellBackgroundByColumnId = <T>(columnId: string, item: T) => {
 export const useCellBackgroundByColumnConfig = <TItem, TItemValue>(
   columnConfig: StandardTableColumnConfig<TItem, TItemValue> | undefined,
   item: TItem
-) => {
+): string | undefined => {
   const { background, backgroundResolver } = columnConfig ?? {};
   return useBackground(backgroundResolver, item, background);
 };